Fix unsigned wrap-around errors

+++ quicksort/quicksort.cpp
@@ -4,16 +4,16 @@
 // https://en.wikipedia.org/wiki/Quicksort#Lomuto_partition_scheme
 
 
-void quicksort_rec(std::vector<long> &A, unsigned long lo, unsigned long hi);
+void quicksort_rec(std::vector<long> &A, long lo, long hi);
 
-unsigned long partition(std::vector<long> &A, unsigned long lo, unsigned long hi);
+long partition(std::vector<long> &A, long lo, long hi);
 
 
 void quicksort(std::vector<long> &A) {
-    quicksort_rec(A, 0, A.size() - 1);
+    quicksort_rec(A, 0, (long)A.size() - 1);
 }
 
-void quicksort_rec(std::vector<long> &A, unsigned long lo, unsigned long hi) {
+void quicksort_rec(std::vector<long> &A, long lo, long hi) {
     if (lo < hi) {
         auto p = partition(A, lo, hi);
         quicksort_rec(A, lo, p - 1);
@@ -21,7 +21,7 @@ void quicksort_rec(std::vector<long> &A, unsigned long lo, unsigned long hi) {
     }
 }
 
-unsigned long partition(std::vector<long> &A, unsigned long lo, unsigned long hi) {
+long partition(std::vector<long> &A, long lo, long hi) {
     auto pivot = A[hi];
     auto i = lo;
     for(auto j = lo; j < hi; ++j) {
